+++ src/components/cluster_detail/cluster_detail_table.js
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types';
 
 import { relativeDate } from '../../lib/helpers.js';
 import AWSAccountID from '../shared/aws_account_id';
-import ReleaseDetailsModal from '../modal/release_details_modal';
+import ReleaseDetailsModal from '../modals/release_details_modal';
 
 class ClusterDetailTable extends React.Component {
   showReleaseDetails = () => {
