+++ src/components/modals/release_details_modal.js
